requirements.txt
alabaster==0.7.12
appnope==0.1.0
arrow==0.15.8
asgiref==3.4.1
asn1crypto==0.24.0
attrs==19.3.0
aws-lambda-builders==1.0.0
aws-sam-cli==1.0.0
aws-sam-translator==1.25.0
awscli
Babel==2.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
boto3==1.13.26
botocore==1.16.26
certifi==2020.4.5.1
cffi==1.12.3
cfn-flip==1.2.1
chardet==3.0.4
charset-normalizer==2.0.6
chevron==0.13.1
Click==7.0
colorama==0.3.9
colorclass==2.2.0
cookiecutter==1.6.0
coverage==5.1
cryptography==3.3.2
cycler==0.10.0
dataclasses==0.8
dateparser==0.7.6
decorator==4.4.0
Django==2.2.24
django-cors-headers==3.10.1
django-environ==0.4.5
django-rest-framework==0.1.0
django-suit==0.2.28
djangorestframework==3.11.0
docker==4.2.2
docopt==0.6.2
docutils==0.14
Flask==1.0.4
future==0.18.2
ghp-import==2.0.1
gunicorn==20.0.4
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==5.2.2
iniconfig==1.1.1
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.9.5
jsonschema==3.2.0
kiwisolver==1.1.0
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mergedeep==1.3.4
mkdocs==1.0.4
more-itertools==8.2.0
numpy==1.16.4
packaging==20.3
parso==0.5.0
pep517==0.11.0
pexpect==4.7.0
pickleshare==0.7.5
pip-tools==6.2.0
pip-upgrader==1.4.15
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.5
pycparser==2.19
pydoc-markdown==2.0.5
Pygments==2.7.4
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.16.0
PySocks==1.7.0
pytest==5.4.1
python-dateutil==2.8.0
python-slugify==5.0.2
pytz==2019.1
PyYAML==5.4
pyyaml_env_tag==0.1
regex==2020.7.14
requests==2.23.0
rsa==4.7
s3transfer==0.3.3
serverlessrepo==0.1.9
six==1.12.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
terminaltables==3.1.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tomlkit==0.5.8
tornado==6.0.3
traitlets==4.3.2
troposphere==2.5.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3
watchdog==2.1.2
wcwidth==0.1.7
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
zipp==3.1.0
